+++ CHANGELOG.md
@@ -9,7 +9,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - added authenticate via API-KEY to webmessage in myems-api
 - added carbon emissions and costs to batch reports of combined equipment, equipment, shopfloor, store and tenant in myems-api and myems-web
 - added Equipment Realtime Monitor to myems-api and myems-web
-- added myems-bacnet to acquire data from BACnet devices
+- added myems-bacnet to acquire data from BACnet devices (HVAC)
+- added myems-s7 to acquire data from S7 devices (PLC)
 
 ### Changed
 - changed carbon unit to KGCO2E in myems-api